added before action and updated show
The before action uses a new method "is user logged in" to check to see if a session id exists before all routes in post except index and show. I updated the create post route logic and added show post route logic so it goes directly to the url. Additionally I updated the post params to include the user_id - twas silly not to. finally, I fixed the post_has_content_and_url? method so it's not always true.